feat(Name, Car): Name 객체 생성 및 Car name field 수정
